requirements.txt
requests
PyYAML
pycryptodome
pypiwin32
easygui
colorama
Brotli
easyyaml
tls-client>=0.2.1